requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Web Framework
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.2
Flask-CORS>=4.0.0

# Database
SQLAlchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
Werkzeug>=2.3.0

# Data Processing
nltk>=3.8.0
emoji>=2.8.0

# Visualization (for admin dashboard)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0

# ASGI Support
asgiref>=3.7.0
